+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.5
+
+- `ASCIIArtTree`:
+  - Added field `pathDelimiter`.
+  - Added `allowGraphs`: support for graphs in the "tree".
+  - Added `totalRoots`, `roots`, `totalLeaves` and `leaves`.
+  - Fix `totalNodes` for graphs.
+  - `_buildTree`: remove recursion to avoid stack overflow and allow big trees.
+  - `toJson`: remove recursion and handles graphs.
+
+- collection: ^1.17.0
+
 ## 1.0.4
 
 - `ASCIIArtTree`: